% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_skim.R
\name{data_skim}
\alias{data_skim}
\title{Brief look at the data before cleaning.}
\usage{
data_skim(data)
}
\arguments{
\item{data}{Dataset as in data frame}
}
\description{
This function returns data frame that shows 
1) features of the input data and 
2) class, number of unique values, and number of NA values for each feature.
The function may be used as a supplement to summary(), str() in basic R package and glimpse() in dplyr package
to gain basic understanding of given dataset before conducting cleaning and processing missing values.
}
\examples{
data_skim(data = df)
}
\keyword{data,}
\keyword{str}
\keyword{summary,}

