% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NofRow.R
\name{NofRow}
\alias{NofRow}
\title{Brief look at the data before processing.}
\usage{
NofRow(data, c)
}
\arguments{
\item{data}{Dataset as in data frame}

\item{c}{index of column that you want to skim through}
}
\description{
This function returns data frame that shows 
1) unique value of a given column of dataset and 
2) its number of occurrence
The function may be used as a supplement to summary(), str() in basic R package and glimpse() in dplyr package
to gain basic understanding of given dataset before conducting cleaning and processing missing values.
}
\examples{
NofRow(data = df, c = column_number)
}
\keyword{data,}
\keyword{str}
\keyword{summary,}

